# api/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": { "description": "User profile retrieved successfully" }
        },
        "tags": ["Auth"]
      },
      "put": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProfileDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "User profile updated successfully" }
        },
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "Eric's Barber API",
    "description": "Auth and Booking API for Barber Shop Application",
    "version": "1.0",
    "contact": {
      "name": "Fahmid Haque",
      "url": "https://www.linkedin.com/in/fahmid-h-b7a96b123/",
      "email": "fahmidulhaque97@pm.me"
    },
    "license": { "name": "MIT", "url": "https://mit-license.org/" }
  },
...
  "servers": [
    {
      "url": "https://erics-barber-api.onrender.com",
      "description": "Base URL for API"
    }
  ],
  "components": {
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "example": "Eric Customer",
            "description": "Display name shown on the user's account"
          }
        },
        "required": ["name"]
      },
...
    }
  }
}